+++ branches/31104branch/src/sitemode/advance.cpp
@@ -326,7 +326,7 @@ void advancecreature(creaturest &cr)
             move(17,1);
             addstr(cr.name);
             addstr("'s wounds.");
-            topmedical->train(SKILL_FIRSTAID,max(50-topmedicalskill*2,0));
+            topmedical->train(SKILL_FIRSTAID,max(int(50-topmedicalskill*2),0));
             cr.wound[w]^=WOUND_BLEEDING;
             refresh();
             getch();
